pyproject.toml
[project]
name = "auto-boxing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.60.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
